+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Pa
 - GeldBetrag & Co in eigenes Modul (wegen Abhängigkeit zu javax.money:money-api)
 - Speichervergleich BigDecimal zu PackedDecimal
 
+## [Unreleased]
+### Added
+- PZN um toShortString() erweitert, der PZN ohne Prefix ausgibt
+
 ## [5.4.3] - 2024-07-09
 ### Fixed
 - fehlende Abhängigkeit von `javax.money:money-api` führt nicht mehr zu fehlerhafter Initialisierung der FachwertFactory-Klasse